refactor(topologies): extract common code into TopologyBase
All three topology types were duplicating code across three files
needlessly.  This reduces duplication, while also making it easier
to maintain and modify in the future by providing a common
TopologyBase class that they all extend.

NODE-1088
