+++ lib/topologies/mongos.js
@@ -1,9 +1,7 @@
 'use strict';
 
-var EventEmitter = require('events').EventEmitter,
-  inherits = require('util').inherits,
-  f = require('util').format,
-  ServerCapabilities = require('./topology_base').ServerCapabilities,
+var ServerCapabilities = require('./topology_base').ServerCapabilities,
+  TopologyBase = require('./topology_base').TopologyBase,
   MongoError = require('mongodb-core').MongoError,
   CMongos = require('mongodb-core').Mongos,
   Cursor = require('../cursor'),
@@ -16,17 +14,7 @@ var EventEmitter = require('events').EventEmitter,
   translateOptions = require('../utils').translateOptions,
   filterOptions = require('../utils').filterOptions,
   mergeOptions = require('../utils').mergeOptions,
-  getReadPreference = require('../utils').getReadPreference,
-  os = require('os'),
   assign = require('../utils').assign;
-
-// Get package.json variable
-var driverVersion = require('../../package.json').version;
-var nodejsversion = f('Node.js %s, %s', process.version, os.endianness());
-var type = os.type();
-var name = process.platform;
-var architecture = process.arch;
-var release = os.release();
 
 /**
  * @fileOverview The **Mongos** class is a class that represents a Mongos Proxy topology and is
@@ -120,144 +108,234 @@ var legalOptionNames = [
  * @property {string} parserType the parser type used (c++ or js).
  * @return {Mongos} a Mongos instance.
  */
-var Mongos = function(servers, options) {
-  if (!(this instanceof Mongos)) return new Mongos(servers, options);
-  options = options || {};
-  var self = this;
+class Mongos extends TopologyBase {
+  constructor(servers, options) {
+    super();
 
-  // Filter the options
-  options = filterOptions(options, legalOptionNames);
+    options = options || {};
+    var self = this;
 
-  // Ensure all the instances are Server
-  for (var i = 0; i < servers.length; i++) {
-    if (!(servers[i] instanceof Server)) {
-      throw MongoError.create({
-        message: 'all seed list instances must be of the Server type',
-        driver: true
+    // Filter the options
+    options = filterOptions(options, legalOptionNames);
+
+    // Ensure all the instances are Server
+    for (var i = 0; i < servers.length; i++) {
+      if (!(servers[i] instanceof Server)) {
+        throw MongoError.create({
+          message: 'all seed list instances must be of the Server type',
+          driver: true
+        });
+      }
+    }
+
+    // Stored options
+    var storeOptions = {
+      force: false,
+      bufferMaxEntries:
+        typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
+    };
+
+    // Shared global store
+    var store = options.store || new Store(self, storeOptions);
+
+    // Build seed list
+    var seedlist = servers.map(function(x) {
+      return { host: x.host, port: x.port };
+    });
+
+    // Get the reconnect option
+    var reconnect = typeof options.auto_reconnect === 'boolean' ? options.auto_reconnect : true;
+    reconnect = typeof options.autoReconnect === 'boolean' ? options.autoReconnect : reconnect;
+
+    // Clone options
+    var clonedOptions = mergeOptions(
+      {},
+      {
+        disconnectHandler: store,
+        cursorFactory: Cursor,
+        reconnect: reconnect,
+        emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
+        size: typeof options.poolSize === 'number' ? options.poolSize : 5
+      }
+    );
+
+    // Translate any SSL options and other connectivity options
+    clonedOptions = translateOptions(clonedOptions, options);
+
+    // Socket options
+    var socketOptions =
+      options.socketOptions && Object.keys(options.socketOptions).length > 0
+        ? options.socketOptions
+        : options;
+
+    // Translate all the options to the mongodb-core ones
+    clonedOptions = translateOptions(clonedOptions, socketOptions);
+    if (typeof clonedOptions.keepAlive === 'number') {
+      clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
+      clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
+    }
+
+    // Build default client information
+    clonedOptions.clientInfo = this.clientInfo;
+    // Do we have an application specific string
+    if (options.appname) {
+      clonedOptions.clientInfo.application = { name: options.appname };
+    }
+
+    // Internal state
+    this.s = {
+      // Create the Mongos
+      coreTopology: new CMongos(seedlist, clonedOptions),
+      // Server capabilities
+      sCapabilities: null,
+      // Debug turned on
+      debug: clonedOptions.debug,
+      // Store option defaults
+      storeOptions: storeOptions,
+      // Cloned options
+      clonedOptions: clonedOptions,
+      // Actual store of callbacks
+      store: store,
+      // Options
+      options: options
+    };
+  }
+
+  // Connect
+  connect(db, _options, callback) {
+    var self = this;
+    if ('function' === typeof _options) (callback = _options), (_options = {});
+    if (_options == null) _options = {};
+    if (!('function' === typeof callback)) callback = null;
+    _options = assign({}, this.s.clonedOptions, _options);
+    self.s.options = _options;
+
+    // Update bufferMaxEntries
+    self.s.storeOptions.bufferMaxEntries = db.bufferMaxEntries;
+
+    // Error handler
+    var connectErrorHandler = function() {
+      return function(err) {
+        // Remove all event handlers
+        var events = ['timeout', 'error', 'close'];
+        events.forEach(function(e) {
+          self.removeListener(e, connectErrorHandler);
+        });
+
+        self.s.coreTopology.removeListener('connect', connectErrorHandler);
+        // Force close the topology
+        self.close(true);
+
+        // Try to callback
+        try {
+          callback(err);
+        } catch (err) {
+          process.nextTick(function() {
+            throw err;
+          });
+        }
+      };
+    };
+
+    // Actual handler
+    var errorHandler = function(event) {
+      return function(err) {
+        if (event !== 'error') {
+          self.emit(event, err);
+        }
+      };
+    };
+
+    // Error handler
+    var reconnectHandler = function() {
+      self.emit('reconnect');
+      self.s.store.execute();
+    };
+
+    // relay the event
+    var relay = function(event) {
+      return function(t, server) {
+        self.emit(event, t, server);
+      };
+    };
+
+    // Connect handler
+    var connectHandler = function() {
+      // Clear out all the current handlers left over
+      var events = ['timeout', 'error', 'close', 'fullsetup'];
+      events.forEach(function(e) {
+        self.s.coreTopology.removeAllListeners(e);
       });
-    }
+
+      // Set up listeners
+      self.s.coreTopology.once('timeout', errorHandler('timeout'));
+      self.s.coreTopology.once('error', errorHandler('error'));
+      self.s.coreTopology.once('close', errorHandler('close'));
+
+      // Set up serverConfig listeners
+      self.s.coreTopology.on('fullsetup', function() {
+        self.emit('fullsetup', self);
+      });
+
+      // Emit open event
+      self.emit('open', null, self);
+
+      // Return correctly
+      try {
+        callback(null, self);
+      } catch (err) {
+        process.nextTick(function() {
+          throw err;
+        });
+      }
+    };
+
+    // Clear out all the current handlers left over
+    var events = [
+      'timeout',
+      'error',
+      'close',
+      'serverOpening',
+      'serverDescriptionChanged',
+      'serverHeartbeatStarted',
+      'serverHeartbeatSucceeded',
+      'serverHeartbeatFailed',
+      'serverClosed',
+      'topologyOpening',
+      'topologyClosed',
+      'topologyDescriptionChanged'
+    ];
+    events.forEach(function(e) {
+      self.s.coreTopology.removeAllListeners(e);
+    });
+
+    // Set up SDAM listeners
+    self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
+    self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
+    self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
+    self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
+    self.s.coreTopology.on('serverOpening', relay('serverOpening'));
+    self.s.coreTopology.on('serverClosed', relay('serverClosed'));
+    self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
+    self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
+    self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
+
+    // Set up listeners
+    self.s.coreTopology.once('timeout', connectErrorHandler('timeout'));
+    self.s.coreTopology.once('error', connectErrorHandler('error'));
+    self.s.coreTopology.once('close', connectErrorHandler('close'));
+    self.s.coreTopology.once('connect', connectHandler);
+    // Join and leave events
+    self.s.coreTopology.on('joined', relay('joined'));
+    self.s.coreTopology.on('left', relay('left'));
+
+    // Reconnect server
+    self.s.coreTopology.on('reconnect', reconnectHandler);
+
+    // Start connection
+    self.s.coreTopology.connect(_options);
   }
-
-  // Stored options
-  var storeOptions = {
-    force: false,
-    bufferMaxEntries:
-      typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
-  };
-
-  // Shared global store
-  var store = options.store || new Store(self, storeOptions);
-
-  // Set up event emitter
-  EventEmitter.call(this);
-
-  // Build seed list
-  var seedlist = servers.map(function(x) {
-    return { host: x.host, port: x.port };
-  });
-
-  // Get the reconnect option
-  var reconnect = typeof options.auto_reconnect === 'boolean' ? options.auto_reconnect : true;
-  reconnect = typeof options.autoReconnect === 'boolean' ? options.autoReconnect : reconnect;
-
-  // Clone options
-  var clonedOptions = mergeOptions(
-    {},
-    {
-      disconnectHandler: store,
-      cursorFactory: Cursor,
-      reconnect: reconnect,
-      emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
-      size: typeof options.poolSize === 'number' ? options.poolSize : 5
-    }
-  );
-
-  // Translate any SSL options and other connectivity options
-  clonedOptions = translateOptions(clonedOptions, options);
-
-  // Socket options
-  var socketOptions =
-    options.socketOptions && Object.keys(options.socketOptions).length > 0
-      ? options.socketOptions
-      : options;
-
-  // Translate all the options to the mongodb-core ones
-  clonedOptions = translateOptions(clonedOptions, socketOptions);
-  if (typeof clonedOptions.keepAlive === 'number') {
-    clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
-    clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
-  }
-
-  // Build default client information
-  this.clientInfo = {
-    driver: {
-      name: 'nodejs',
-      version: driverVersion
-    },
-    os: {
-      type: type,
-      name: name,
-      architecture: architecture,
-      version: release
-    },
-    platform: nodejsversion
-  };
-
-  // Build default client information
-  clonedOptions.clientInfo = this.clientInfo;
-  // Do we have an application specific string
-  if (options.appname) {
-    clonedOptions.clientInfo.application = { name: options.appname };
-  }
-
-  // Internal state
-  this.s = {
-    // Create the Mongos
-    coreTopology: new CMongos(seedlist, clonedOptions),
-    // Server capabilities
-    sCapabilities: null,
-    // Debug turned on
-    debug: clonedOptions.debug,
-    // Store option defaults
-    storeOptions: storeOptions,
-    // Cloned options
-    clonedOptions: clonedOptions,
-    // Actual store of callbacks
-    store: store,
-    // Options
-    options: options
-  };
-};
-
-var define = (Mongos.define = new Define('Mongos', Mongos, false));
-
-/**
- * @ignore
- */
-inherits(Mongos, EventEmitter);
-
-// Last ismaster
-Object.defineProperty(Mongos.prototype, 'isMasterDoc', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.lastIsMaster();
-  }
-});
-
-Object.defineProperty(Mongos.prototype, 'parserType', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.parserType;
-  }
-});
-
-// BSON property
-Object.defineProperty(Mongos.prototype, 'bson', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.s.bson;
-  }
-});
+}
 
 Object.defineProperty(Mongos.prototype, 'haInterval', {
   enumerable: true,
@@ -266,264 +344,28 @@ Object.defineProperty(Mongos.prototype, 'haInterval', {
   }
 });
 
-Object.defineProperty(Mongos.prototype, 'logicalSessionTimeoutMinutes', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.logicalSessionTimeoutMinutes;
-  }
-});
-
-// Connect
-Mongos.prototype.connect = function(db, _options, callback) {
-  var self = this;
-  if ('function' === typeof _options) (callback = _options), (_options = {});
-  if (_options == null) _options = {};
-  if (!('function' === typeof callback)) callback = null;
-  _options = assign({}, this.s.clonedOptions, _options);
-  self.s.options = _options;
-
-  // Update bufferMaxEntries
-  self.s.storeOptions.bufferMaxEntries = db.bufferMaxEntries;
-
-  // Error handler
-  var connectErrorHandler = function() {
-    return function(err) {
-      // Remove all event handlers
-      var events = ['timeout', 'error', 'close'];
-      events.forEach(function(e) {
-        self.removeListener(e, connectErrorHandler);
-      });
-
-      self.s.coreTopology.removeListener('connect', connectErrorHandler);
-      // Force close the topology
-      self.close(true);
-
-      // Try to callback
-      try {
-        callback(err);
-      } catch (err) {
-        process.nextTick(function() {
-          throw err;
-        });
-      }
-    };
-  };
-
-  // Actual handler
-  var errorHandler = function(event) {
-    return function(err) {
-      if (event !== 'error') {
-        self.emit(event, err);
-      }
-    };
-  };
-
-  // Error handler
-  var reconnectHandler = function() {
-    self.emit('reconnect');
-    self.s.store.execute();
-  };
-
-  // relay the event
-  var relay = function(event) {
-    return function(t, server) {
-      self.emit(event, t, server);
-    };
-  };
-
-  // Connect handler
-  var connectHandler = function() {
-    // Clear out all the current handlers left over
-    var events = ['timeout', 'error', 'close', 'fullsetup'];
-    events.forEach(function(e) {
-      self.s.coreTopology.removeAllListeners(e);
-    });
-
-    // Set up listeners
-    self.s.coreTopology.once('timeout', errorHandler('timeout'));
-    self.s.coreTopology.once('error', errorHandler('error'));
-    self.s.coreTopology.once('close', errorHandler('close'));
-
-    // Set up serverConfig listeners
-    self.s.coreTopology.on('fullsetup', function() {
-      self.emit('fullsetup', self);
-    });
-
-    // Emit open event
-    self.emit('open', null, self);
-
-    // Return correctly
-    try {
-      callback(null, self);
-    } catch (err) {
-      process.nextTick(function() {
-        throw err;
-      });
-    }
-  };
-
-  // Clear out all the current handlers left over
-  var events = [
-    'timeout',
-    'error',
-    'close',
-    'serverOpening',
-    'serverDescriptionChanged',
-    'serverHeartbeatStarted',
-    'serverHeartbeatSucceeded',
-    'serverHeartbeatFailed',
-    'serverClosed',
-    'topologyOpening',
-    'topologyClosed',
-    'topologyDescriptionChanged'
-  ];
-  events.forEach(function(e) {
-    self.s.coreTopology.removeAllListeners(e);
-  });
-
-  // Set up SDAM listeners
-  self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
-  self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
-  self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
-  self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
-  self.s.coreTopology.on('serverOpening', relay('serverOpening'));
-  self.s.coreTopology.on('serverClosed', relay('serverClosed'));
-  self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
-  self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
-  self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
-
-  // Set up listeners
-  self.s.coreTopology.once('timeout', connectErrorHandler('timeout'));
-  self.s.coreTopology.once('error', connectErrorHandler('error'));
-  self.s.coreTopology.once('close', connectErrorHandler('close'));
-  self.s.coreTopology.once('connect', connectHandler);
-  // Join and leave events
-  self.s.coreTopology.on('joined', relay('joined'));
-  self.s.coreTopology.on('left', relay('left'));
-
-  // Reconnect server
-  self.s.coreTopology.on('reconnect', reconnectHandler);
-
-  // Start connection
-  self.s.coreTopology.connect(_options);
-};
-
-// Server capabilities
-Mongos.prototype.capabilities = function() {
-  if (this.s.sCapabilities) return this.s.sCapabilities;
-  if (this.s.coreTopology.lastIsMaster() == null) return null;
-  this.s.sCapabilities = new ServerCapabilities(this.s.coreTopology.lastIsMaster());
-  return this.s.sCapabilities;
-};
-
+const define = (Mongos.define = new Define('Mongos', Mongos, false));
 define.classMethod('capabilities', {
   callback: false,
   promise: false,
   returns: [ServerCapabilities]
 });
 
-// Command
-Mongos.prototype.command = function(ns, cmd, options, callback) {
-  this.s.coreTopology.command(ns, cmd, getReadPreference(options), callback);
-};
-
 define.classMethod('command', { callback: true, promise: false });
-
-// Insert
-Mongos.prototype.insert = function(ns, ops, options, callback) {
-  this.s.coreTopology.insert(ns, ops, options, function(e, m) {
-    callback(e, m);
-  });
-};
-
 define.classMethod('insert', { callback: true, promise: false });
-
-// Update
-Mongos.prototype.update = function(ns, ops, options, callback) {
-  this.s.coreTopology.update(ns, ops, options, callback);
-};
-
 define.classMethod('update', { callback: true, promise: false });
-
-// Remove
-Mongos.prototype.remove = function(ns, ops, options, callback) {
-  this.s.coreTopology.remove(ns, ops, options, callback);
-};
-
 define.classMethod('remove', { callback: true, promise: false });
-
-// Destroyed
-Mongos.prototype.isDestroyed = function() {
-  return this.s.coreTopology.isDestroyed();
-};
-
-// IsConnected
-Mongos.prototype.isConnected = function() {
-  return this.s.coreTopology.isConnected();
-};
-
 define.classMethod('isConnected', { callback: false, promise: false, returns: [Boolean] });
-
-// Insert
-Mongos.prototype.cursor = function(ns, cmd, options) {
-  options.disconnectHandler = this.s.store;
-  return this.s.coreTopology.cursor(ns, cmd, options);
-};
-
+define.classMethod('isDestroyed', { callback: false, promise: false, returns: [Boolean] });
 define.classMethod('cursor', {
   callback: false,
   promise: false,
   returns: [Cursor, AggregationCursor, CommandCursor]
 });
 
-Mongos.prototype.lastIsMaster = function() {
-  return this.s.coreTopology.lastIsMaster();
-};
-
-/**
- * Unref all sockets
- * @method
- */
-Mongos.prototype.unref = function() {
-  return this.s.coreTopology.unref();
-};
-
-Mongos.prototype.close = function(forceClosed) {
-  this.s.coreTopology.destroy({
-    force: typeof forceClosed === 'boolean' ? forceClosed : false
-  });
-  // We need to wash out all stored processes
-  if (forceClosed === true) {
-    this.s.storeOptions.force = forceClosed;
-    this.s.store.flush();
-  }
-};
-
 define.classMethod('close', { callback: false, promise: false });
-
-Mongos.prototype.auth = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.auth.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('auth', { callback: true, promise: false });
-
-Mongos.prototype.logout = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.logout.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('logout', { callback: true, promise: false });
-
-/**
- * All raw connections
- * @method
- * @return {array}
- */
-Mongos.prototype.connections = function() {
-  return this.s.coreTopology.connections();
-};
-
 define.classMethod('connections', { callback: false, promise: false, returns: [Array] });
 
 /**
