+++ lib/topologies/replset.js
@@ -1,25 +1,19 @@
 'use strict';
 
-var EventEmitter = require('events').EventEmitter,
-  inherits = require('util').inherits,
-  f = require('util').format,
-  Server = require('./server'),
+var Server = require('./server'),
   Cursor = require('../cursor'),
   AggregationCursor = require('../aggregation_cursor'),
   CommandCursor = require('../command_cursor'),
-  ReadPreference = require('../read_preference'),
   MongoError = require('mongodb-core').MongoError,
   ServerCapabilities = require('./topology_base').ServerCapabilities,
+  TopologyBase = require('./topology_base').TopologyBase,
   Store = require('./topology_base').Store,
   Define = require('../metadata'),
   CReplSet = require('mongodb-core').ReplSet,
-  CoreReadPreference = require('mongodb-core').ReadPreference,
   MAX_JS_INT = require('../utils').MAX_JS_INT,
   translateOptions = require('../utils').translateOptions,
   filterOptions = require('../utils').filterOptions,
-  getReadPreference = require('../utils').getReadPreference,
   mergeOptions = require('../utils').mergeOptions,
-  os = require('os'),
   assign = require('../utils').assign;
 
 /**
@@ -83,14 +77,6 @@ var legalOptionNames = [
   'maxStalenessSeconds'
 ];
 
-// Get package.json variable
-var driverVersion = require('../../package.json').version;
-var nodejsversion = f('Node.js %s, %s', process.version, os.endianness());
-var type = os.type();
-var name = process.platform;
-var architecture = process.arch;
-var release = os.release();
-
 /**
  * Creates a new ReplSet instance
  * @class
@@ -132,157 +118,281 @@ var release = os.release();
  * @property {string} parserType the parser type used (c++ or js).
  * @return {ReplSet} a ReplSet instance.
  */
-var ReplSet = function(servers, options) {
-  if (!(this instanceof ReplSet)) return new ReplSet(servers, options);
-  options = options || {};
-  var self = this;
-  // Set up event emitter
-  EventEmitter.call(this);
+class ReplSet extends TopologyBase {
+  constructor(servers, options) {
+    super();
 
-  // Filter the options
-  options = filterOptions(options, legalOptionNames);
+    options = options || {};
+    var self = this;
 
-  // Ensure all the instances are Server
-  for (var i = 0; i < servers.length; i++) {
-    if (!(servers[i] instanceof Server)) {
-      throw MongoError.create({
-        message: 'all seed list instances must be of the Server type',
-        driver: true
+    // Filter the options
+    options = filterOptions(options, legalOptionNames);
+
+    // Ensure all the instances are Server
+    for (var i = 0; i < servers.length; i++) {
+      if (!(servers[i] instanceof Server)) {
+        throw MongoError.create({
+          message: 'all seed list instances must be of the Server type',
+          driver: true
+        });
+      }
+    }
+
+    // Stored options
+    var storeOptions = {
+      force: false,
+      bufferMaxEntries:
+        typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
+    };
+
+    // Shared global store
+    var store = options.store || new Store(self, storeOptions);
+
+    // Build seed list
+    var seedlist = servers.map(function(x) {
+      return { host: x.host, port: x.port };
+    });
+
+    // Clone options
+    var clonedOptions = mergeOptions(
+      {},
+      {
+        disconnectHandler: store,
+        cursorFactory: Cursor,
+        reconnect: false,
+        emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
+        size: typeof options.poolSize === 'number' ? options.poolSize : 5
+      }
+    );
+
+    // Translate any SSL options and other connectivity options
+    clonedOptions = translateOptions(clonedOptions, options);
+
+    // Socket options
+    var socketOptions =
+      options.socketOptions && Object.keys(options.socketOptions).length > 0
+        ? options.socketOptions
+        : options;
+
+    // Translate all the options to the mongodb-core ones
+    clonedOptions = translateOptions(clonedOptions, socketOptions);
+    if (typeof clonedOptions.keepAlive === 'number') {
+      clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
+      clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
+    }
+
+    // Build default client information
+    clonedOptions.clientInfo = this.clientInfo;
+    // Do we have an application specific string
+    if (options.appname) {
+      clonedOptions.clientInfo.application = { name: options.appname };
+    }
+
+    // Create the ReplSet
+    var coreTopology = new CReplSet(seedlist, clonedOptions);
+
+    // Listen to reconnect event
+    coreTopology.on('reconnect', function() {
+      self.emit('reconnect');
+      store.execute();
+    });
+
+    // Internal state
+    this.s = {
+      // Replicaset
+      coreTopology: coreTopology,
+      // Server capabilities
+      sCapabilities: null,
+      // Debug tag
+      tag: options.tag,
+      // Store options
+      storeOptions: storeOptions,
+      // Cloned options
+      clonedOptions: clonedOptions,
+      // Store
+      store: store,
+      // Options
+      options: options
+    };
+
+    // Debug
+    if (clonedOptions.debug) {
+      // Last ismaster
+      Object.defineProperty(this, 'replset', {
+        enumerable: true,
+        get: function() {
+          return coreTopology;
+        }
       });
     }
   }
 
-  // Stored options
-  var storeOptions = {
-    force: false,
-    bufferMaxEntries:
-      typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
-  };
+  // Connect method
+  connect(db, _options, callback) {
+    var self = this;
+    if ('function' === typeof _options) (callback = _options), (_options = {});
+    if (_options == null) _options = {};
+    if (!('function' === typeof callback)) callback = null;
+    _options = assign({}, this.s.clonedOptions, _options);
+    self.s.options = _options;
 
-  // Shared global store
-  var store = options.store || new Store(self, storeOptions);
+    // Update bufferMaxEntries
+    self.s.storeOptions.bufferMaxEntries = db.bufferMaxEntries;
 
-  // Build seed list
-  var seedlist = servers.map(function(x) {
-    return { host: x.host, port: x.port };
-  });
+    // Actual handler
+    var errorHandler = function(event) {
+      return function(err) {
+        if (event !== 'error') {
+          self.emit(event, err);
+        }
+      };
+    };
 
-  // Clone options
-  var clonedOptions = mergeOptions(
-    {},
-    {
-      disconnectHandler: store,
-      cursorFactory: Cursor,
-      reconnect: false,
-      emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
-      size: typeof options.poolSize === 'number' ? options.poolSize : 5
-    }
-  );
+    // Clear out all the current handlers left over
+    var events = [
+      'timeout',
+      'error',
+      'close',
+      'serverOpening',
+      'serverDescriptionChanged',
+      'serverHeartbeatStarted',
+      'serverHeartbeatSucceeded',
+      'serverHeartbeatFailed',
+      'serverClosed',
+      'topologyOpening',
+      'topologyClosed',
+      'topologyDescriptionChanged',
+      'joined',
+      'left',
+      'ping',
+      'ha'
+    ];
+    events.forEach(function(e) {
+      self.s.coreTopology.removeAllListeners(e);
+    });
 
-  // Translate any SSL options and other connectivity options
-  clonedOptions = translateOptions(clonedOptions, options);
+    // relay the event
+    var relay = function(event) {
+      return function(t, server) {
+        self.emit(event, t, server);
+      };
+    };
 
-  // Socket options
-  var socketOptions =
-    options.socketOptions && Object.keys(options.socketOptions).length > 0
-      ? options.socketOptions
-      : options;
+    // Replset events relay
+    var replsetRelay = function(event) {
+      return function(t, server) {
+        self.emit(event, t, server.lastIsMaster(), server);
+      };
+    };
 
-  // Translate all the options to the mongodb-core ones
-  clonedOptions = translateOptions(clonedOptions, socketOptions);
-  if (typeof clonedOptions.keepAlive === 'number') {
-    clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
-    clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
-  }
+    // Relay ha
+    var relayHa = function(t, state) {
+      self.emit('ha', t, state);
 
-  // Client info
-  this.clientInfo = {
-    driver: {
-      name: 'nodejs',
-      version: driverVersion
-    },
-    os: {
-      type: type,
-      name: name,
-      architecture: architecture,
-      version: release
-    },
-    platform: nodejsversion
-  };
-
-  // Build default client information
-  clonedOptions.clientInfo = this.clientInfo;
-  // Do we have an application specific string
-  if (options.appname) {
-    clonedOptions.clientInfo.application = { name: options.appname };
-  }
-
-  // Create the ReplSet
-  var coreTopology = new CReplSet(seedlist, clonedOptions);
-
-  // Listen to reconnect event
-  coreTopology.on('reconnect', function() {
-    self.emit('reconnect');
-    store.execute();
-  });
-
-  // Internal state
-  this.s = {
-    // Replicaset
-    coreTopology: coreTopology,
-    // Server capabilities
-    sCapabilities: null,
-    // Debug tag
-    tag: options.tag,
-    // Store options
-    storeOptions: storeOptions,
-    // Cloned options
-    clonedOptions: clonedOptions,
-    // Store
-    store: store,
-    // Options
-    options: options
-  };
-
-  // Debug
-  if (clonedOptions.debug) {
-    // Last ismaster
-    Object.defineProperty(this, 'replset', {
-      enumerable: true,
-      get: function() {
-        return coreTopology;
+      if (t === 'start') {
+        self.emit('ha_connect', t, state);
+      } else if (t === 'end') {
+        self.emit('ha_ismaster', t, state);
       }
+    };
+
+    // Set up serverConfig listeners
+    self.s.coreTopology.on('joined', replsetRelay('joined'));
+    self.s.coreTopology.on('left', relay('left'));
+    self.s.coreTopology.on('ping', relay('ping'));
+    self.s.coreTopology.on('ha', relayHa);
+
+    // Set up SDAM listeners
+    self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
+    self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
+    self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
+    self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
+    self.s.coreTopology.on('serverOpening', relay('serverOpening'));
+    self.s.coreTopology.on('serverClosed', relay('serverClosed'));
+    self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
+    self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
+    self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
+
+    self.s.coreTopology.on('fullsetup', function() {
+      self.emit('fullsetup', self, self);
+    });
+
+    self.s.coreTopology.on('all', function() {
+      self.emit('all', null, self);
+    });
+
+    // Connect handler
+    var connectHandler = function() {
+      // Set up listeners
+      self.s.coreTopology.once('timeout', errorHandler('timeout'));
+      self.s.coreTopology.once('error', errorHandler('error'));
+      self.s.coreTopology.once('close', errorHandler('close'));
+
+      // Emit open event
+      self.emit('open', null, self);
+
+      // Return correctly
+      try {
+        callback(null, self);
+      } catch (err) {
+        process.nextTick(function() {
+          throw err;
+        });
+      }
+    };
+
+    // Error handler
+    var connectErrorHandler = function() {
+      return function(err) {
+        ['timeout', 'error', 'close'].forEach(function(e) {
+          self.s.coreTopology.removeListener(e, connectErrorHandler);
+        });
+
+        self.s.coreTopology.removeListener('connect', connectErrorHandler);
+        // Destroy the replset
+        self.s.coreTopology.destroy();
+
+        // Try to callback
+        try {
+          callback(err);
+        } catch (err) {
+          if (!self.s.coreTopology.isConnected())
+            process.nextTick(function() {
+              throw err;
+            });
+        }
+      };
+    };
+
+    // Set up listeners
+    self.s.coreTopology.once('timeout', connectErrorHandler('timeout'));
+    self.s.coreTopology.once('error', connectErrorHandler('error'));
+    self.s.coreTopology.once('close', connectErrorHandler('close'));
+    self.s.coreTopology.once('connect', connectHandler);
+
+    // Start connection
+    self.s.coreTopology.connect(_options);
+  }
+
+  close(forceClosed) {
+    var self = this;
+    // Call destroy on the topology
+    this.s.coreTopology.destroy({
+      force: typeof forceClosed === 'boolean' ? forceClosed : false
+    });
+
+    // We need to wash out all stored processes
+    if (forceClosed === true) {
+      this.s.storeOptions.force = forceClosed;
+      this.s.store.flush();
+    }
+
+    var events = ['timeout', 'error', 'close', 'joined', 'left'];
+    events.forEach(function(e) {
+      self.removeAllListeners(e);
     });
   }
-};
-
-/**
- * @ignore
- */
-inherits(ReplSet, EventEmitter);
-
-// Last ismaster
-Object.defineProperty(ReplSet.prototype, 'isMasterDoc', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.lastIsMaster();
-  }
-});
-
-Object.defineProperty(ReplSet.prototype, 'parserType', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.parserType;
-  }
-});
-
-// BSON property
-Object.defineProperty(ReplSet.prototype, 'bson', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.s.bson;
-  }
-});
+}
 
 Object.defineProperty(ReplSet.prototype, 'haInterval', {
   enumerable: true,
@@ -291,307 +401,27 @@ Object.defineProperty(ReplSet.prototype, 'haInterval', {
   }
 });
 
-Object.defineProperty(ReplSet.prototype, 'logicalSessionTimeoutMinutes', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.logicalSessionTimeoutMinutes;
-  }
-});
-
-var define = (ReplSet.define = new Define('ReplSet', ReplSet, false));
-
-// Ensure the right read Preference object
-var translateReadPreference = function(options) {
-  if (typeof options.readPreference === 'string') {
-    options.readPreference = new CoreReadPreference(options.readPreference);
-  } else if (options.readPreference instanceof ReadPreference) {
-    options.readPreference = new CoreReadPreference(
-      options.readPreference.mode,
-      options.readPreference.tags,
-      { maxStalenessSeconds: options.readPreference.maxStalenessSeconds }
-    );
-  }
-
-  return options;
-};
-
-// Connect method
-ReplSet.prototype.connect = function(db, _options, callback) {
-  var self = this;
-  if ('function' === typeof _options) (callback = _options), (_options = {});
-  if (_options == null) _options = {};
-  if (!('function' === typeof callback)) callback = null;
-  _options = assign({}, this.s.clonedOptions, _options);
-  self.s.options = _options;
-
-  // Update bufferMaxEntries
-  self.s.storeOptions.bufferMaxEntries = db.bufferMaxEntries;
-
-  // Actual handler
-  var errorHandler = function(event) {
-    return function(err) {
-      if (event !== 'error') {
-        self.emit(event, err);
-      }
-    };
-  };
-
-  // Clear out all the current handlers left over
-  var events = [
-    'timeout',
-    'error',
-    'close',
-    'serverOpening',
-    'serverDescriptionChanged',
-    'serverHeartbeatStarted',
-    'serverHeartbeatSucceeded',
-    'serverHeartbeatFailed',
-    'serverClosed',
-    'topologyOpening',
-    'topologyClosed',
-    'topologyDescriptionChanged',
-    'joined',
-    'left',
-    'ping',
-    'ha'
-  ];
-  events.forEach(function(e) {
-    self.s.coreTopology.removeAllListeners(e);
-  });
-
-  // relay the event
-  var relay = function(event) {
-    return function(t, server) {
-      self.emit(event, t, server);
-    };
-  };
-
-  // Replset events relay
-  var replsetRelay = function(event) {
-    return function(t, server) {
-      self.emit(event, t, server.lastIsMaster(), server);
-    };
-  };
-
-  // Relay ha
-  var relayHa = function(t, state) {
-    self.emit('ha', t, state);
-
-    if (t === 'start') {
-      self.emit('ha_connect', t, state);
-    } else if (t === 'end') {
-      self.emit('ha_ismaster', t, state);
-    }
-  };
-
-  // Set up serverConfig listeners
-  self.s.coreTopology.on('joined', replsetRelay('joined'));
-  self.s.coreTopology.on('left', relay('left'));
-  self.s.coreTopology.on('ping', relay('ping'));
-  self.s.coreTopology.on('ha', relayHa);
-
-  // Set up SDAM listeners
-  self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
-  self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
-  self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
-  self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
-  self.s.coreTopology.on('serverOpening', relay('serverOpening'));
-  self.s.coreTopology.on('serverClosed', relay('serverClosed'));
-  self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
-  self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
-  self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
-
-  self.s.coreTopology.on('fullsetup', function() {
-    self.emit('fullsetup', self, self);
-  });
-
-  self.s.coreTopology.on('all', function() {
-    self.emit('all', null, self);
-  });
-
-  // Connect handler
-  var connectHandler = function() {
-    // Set up listeners
-    self.s.coreTopology.once('timeout', errorHandler('timeout'));
-    self.s.coreTopology.once('error', errorHandler('error'));
-    self.s.coreTopology.once('close', errorHandler('close'));
-
-    // Emit open event
-    self.emit('open', null, self);
-
-    // Return correctly
-    try {
-      callback(null, self);
-    } catch (err) {
-      process.nextTick(function() {
-        throw err;
-      });
-    }
-  };
-
-  // Error handler
-  var connectErrorHandler = function() {
-    return function(err) {
-      ['timeout', 'error', 'close'].forEach(function(e) {
-        self.s.coreTopology.removeListener(e, connectErrorHandler);
-      });
-
-      self.s.coreTopology.removeListener('connect', connectErrorHandler);
-      // Destroy the replset
-      self.s.coreTopology.destroy();
-
-      // Try to callback
-      try {
-        callback(err);
-      } catch (err) {
-        if (!self.s.coreTopology.isConnected())
-          process.nextTick(function() {
-            throw err;
-          });
-      }
-    };
-  };
-
-  // Set up listeners
-  self.s.coreTopology.once('timeout', connectErrorHandler('timeout'));
-  self.s.coreTopology.once('error', connectErrorHandler('error'));
-  self.s.coreTopology.once('close', connectErrorHandler('close'));
-  self.s.coreTopology.once('connect', connectHandler);
-
-  // Start connection
-  self.s.coreTopology.connect(_options);
-};
-
-// Server capabilities
-ReplSet.prototype.capabilities = function() {
-  if (this.s.sCapabilities) return this.s.sCapabilities;
-  if (this.s.coreTopology.lastIsMaster() == null) return null;
-  this.s.sCapabilities = new ServerCapabilities(this.s.coreTopology.lastIsMaster());
-  return this.s.sCapabilities;
-};
-
+const define = (ReplSet.define = new Define('ReplSet', ReplSet, false));
 define.classMethod('capabilities', {
   callback: false,
   promise: false,
   returns: [ServerCapabilities]
 });
 
-// Command
-ReplSet.prototype.command = function(ns, cmd, options, callback) {
-  this.s.coreTopology.command(ns, cmd, getReadPreference(options), callback);
-};
-
 define.classMethod('command', { callback: true, promise: false });
-
-// Insert
-ReplSet.prototype.insert = function(ns, ops, options, callback) {
-  this.s.coreTopology.insert(ns, ops, options, callback);
-};
-
 define.classMethod('insert', { callback: true, promise: false });
-
-// Update
-ReplSet.prototype.update = function(ns, ops, options, callback) {
-  this.s.coreTopology.update(ns, ops, options, callback);
-};
-
 define.classMethod('update', { callback: true, promise: false });
-
-// Remove
-ReplSet.prototype.remove = function(ns, ops, options, callback) {
-  this.s.coreTopology.remove(ns, ops, options, callback);
-};
-
 define.classMethod('remove', { callback: true, promise: false });
-
-// Destroyed
-ReplSet.prototype.isDestroyed = function() {
-  return this.s.coreTopology.isDestroyed();
-};
-
-// IsConnected
-ReplSet.prototype.isConnected = function(options) {
-  options = options || {};
-
-  // If we passed in a readPreference, translate to
-  // a CoreReadPreference instance
-  if (options.readPreference) {
-    options.readPreference = translateReadPreference(options.readPreference);
-  }
-
-  return this.s.coreTopology.isConnected(options);
-};
-
 define.classMethod('isConnected', { callback: false, promise: false, returns: [Boolean] });
-
-// Insert
-ReplSet.prototype.cursor = function(ns, cmd, options) {
-  options = translateReadPreference(options);
-  options.disconnectHandler = this.s.store;
-  return this.s.coreTopology.cursor(ns, cmd, options);
-};
-
 define.classMethod('cursor', {
   callback: false,
   promise: false,
   returns: [Cursor, AggregationCursor, CommandCursor]
 });
 
-ReplSet.prototype.lastIsMaster = function() {
-  return this.s.coreTopology.lastIsMaster();
-};
-
-/**
- * Unref all sockets
- * @method
- */
-ReplSet.prototype.unref = function() {
-  return this.s.coreTopology.unref();
-};
-
-ReplSet.prototype.close = function(forceClosed) {
-  var self = this;
-  // Call destroy on the topology
-  this.s.coreTopology.destroy({
-    force: typeof forceClosed === 'boolean' ? forceClosed : false
-  });
-  // We need to wash out all stored processes
-  if (forceClosed === true) {
-    this.s.storeOptions.force = forceClosed;
-    this.s.store.flush();
-  }
-
-  var events = ['timeout', 'error', 'close', 'joined', 'left'];
-  events.forEach(function(e) {
-    self.removeAllListeners(e);
-  });
-};
-
 define.classMethod('close', { callback: false, promise: false });
-
-ReplSet.prototype.auth = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.auth.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('auth', { callback: true, promise: false });
-
-ReplSet.prototype.logout = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.logout.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('logout', { callback: true, promise: false });
-
-/**
- * All raw connections
- * @method
- * @return {array}
- */
-ReplSet.prototype.connections = function() {
-  return this.s.coreTopology.connections();
-};
-
 define.classMethod('connections', { callback: false, promise: false, returns: [Array] });
 
 /**
