+++ lib/topologies/server.js
@@ -1,13 +1,11 @@
 'use strict';
 
-var EventEmitter = require('events').EventEmitter,
-  inherits = require('util').inherits,
-  CServer = require('mongodb-core').Server,
+var CServer = require('mongodb-core').Server,
   Cursor = require('../cursor'),
   AggregationCursor = require('../aggregation_cursor'),
   CommandCursor = require('../command_cursor'),
-  f = require('util').format,
   ServerCapabilities = require('./topology_base').ServerCapabilities,
+  TopologyBase = require('./topology_base').TopologyBase,
   Store = require('./topology_base').Store,
   Define = require('../metadata'),
   MongoError = require('mongodb-core').MongoError,
@@ -15,17 +13,7 @@ var EventEmitter = require('events').EventEmitter,
   translateOptions = require('../utils').translateOptions,
   filterOptions = require('../utils').filterOptions,
   mergeOptions = require('../utils').mergeOptions,
-  getReadPreference = require('../utils').getReadPreference,
-  os = require('os'),
   assign = require('../utils').assign;
-
-// Get package.json variable
-var driverVersion = require('../../package.json').version;
-var nodejsversion = f('Node.js %s, %s', process.version, os.endianness());
-var type = os.type();
-var name = process.platform;
-var architecture = process.arch;
-var release = os.release();
 
 /**
  * @fileOverview The **Server** class is a class that represents a single server topology and is
@@ -120,146 +108,250 @@ var legalOptionNames = [
  * @property {string} parserType the parser type used (c++ or js).
  * @return {Server} a Server instance.
  */
-var Server = function(host, port, options) {
-  options = options || {};
-  if (!(this instanceof Server)) return new Server(host, port, options);
-  EventEmitter.call(this);
-  var self = this;
+class Server extends TopologyBase {
+  constructor(host, port, options) {
+    super();
+    var self = this;
 
-  // Filter the options
-  options = filterOptions(options, legalOptionNames);
+    // Filter the options
+    options = filterOptions(options, legalOptionNames);
 
-  // Stored options
-  var storeOptions = {
-    force: false,
-    bufferMaxEntries:
-      typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
-  };
+    // Stored options
+    var storeOptions = {
+      force: false,
+      bufferMaxEntries:
+        typeof options.bufferMaxEntries === 'number' ? options.bufferMaxEntries : MAX_JS_INT
+    };
 
-  // Shared global store
-  var store = options.store || new Store(self, storeOptions);
+    // Shared global store
+    var store = options.store || new Store(self, storeOptions);
 
-  // Detect if we have a socket connection
-  if (host.indexOf('/') !== -1) {
-    if (port != null && typeof port === 'object') {
-      options = port;
-      port = null;
+    // Detect if we have a socket connection
+    if (host.indexOf('/') !== -1) {
+      if (port != null && typeof port === 'object') {
+        options = port;
+        port = null;
+      }
+    } else if (port == null) {
+      throw MongoError.create({ message: 'port must be specified', driver: true });
     }
-  } else if (port == null) {
-    throw MongoError.create({ message: 'port must be specified', driver: true });
-  }
 
-  // Get the reconnect option
-  var reconnect = typeof options.auto_reconnect === 'boolean' ? options.auto_reconnect : true;
-  reconnect = typeof options.autoReconnect === 'boolean' ? options.autoReconnect : reconnect;
+    // Get the reconnect option
+    var reconnect = typeof options.auto_reconnect === 'boolean' ? options.auto_reconnect : true;
+    reconnect = typeof options.autoReconnect === 'boolean' ? options.autoReconnect : reconnect;
 
-  // Clone options
-  var clonedOptions = mergeOptions(
-    {},
-    {
+    // Clone options
+    var clonedOptions = mergeOptions(
+      {},
+      {
+        host: host,
+        port: port,
+        disconnectHandler: store,
+        cursorFactory: Cursor,
+        reconnect: reconnect,
+        emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
+        size: typeof options.poolSize === 'number' ? options.poolSize : 5
+      }
+    );
+
+    // Translate any SSL options and other connectivity options
+    clonedOptions = translateOptions(clonedOptions, options);
+
+    // Socket options
+    var socketOptions =
+      options.socketOptions && Object.keys(options.socketOptions).length > 0
+        ? options.socketOptions
+        : options;
+
+    // Translate all the options to the mongodb-core ones
+    clonedOptions = translateOptions(clonedOptions, socketOptions);
+    if (typeof clonedOptions.keepAlive === 'number') {
+      clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
+      clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
+    }
+
+    // Build default client information
+    clonedOptions.clientInfo = this.clientInfo;
+    // Do we have an application specific string
+    if (options.appname) {
+      clonedOptions.clientInfo.application = { name: options.appname };
+    }
+
+    // Define the internal properties
+    this.s = {
+      // Create an instance of a server instance from mongodb-core
+      coreTopology: new CServer(clonedOptions),
+      // Server capabilities
+      sCapabilities: null,
+      // Cloned options
+      clonedOptions: clonedOptions,
+      // Reconnect
+      reconnect: clonedOptions.reconnect,
+      // Emit error
+      emitError: clonedOptions.emitError,
+      // Pool size
+      poolSize: clonedOptions.size,
+      // Store Options
+      storeOptions: storeOptions,
+      // Store
+      store: store,
+      // Host
       host: host,
+      // Port
       port: port,
-      disconnectHandler: store,
-      cursorFactory: Cursor,
-      reconnect: reconnect,
-      emitError: typeof options.emitError === 'boolean' ? options.emitError : true,
-      size: typeof options.poolSize === 'number' ? options.poolSize : 5
-    }
-  );
-
-  // Translate any SSL options and other connectivity options
-  clonedOptions = translateOptions(clonedOptions, options);
-
-  // Socket options
-  var socketOptions =
-    options.socketOptions && Object.keys(options.socketOptions).length > 0
-      ? options.socketOptions
-      : options;
-
-  // Translate all the options to the mongodb-core ones
-  clonedOptions = translateOptions(clonedOptions, socketOptions);
-  if (typeof clonedOptions.keepAlive === 'number') {
-    clonedOptions.keepAliveInitialDelay = clonedOptions.keepAlive;
-    clonedOptions.keepAlive = clonedOptions.keepAlive > 0;
+      // Options
+      options: options
+    };
   }
 
-  // Build default client information
-  this.clientInfo = {
-    driver: {
-      name: 'nodejs',
-      version: driverVersion
-    },
-    os: {
-      type: type,
-      name: name,
-      architecture: architecture,
-      version: release
-    },
-    platform: nodejsversion
-  };
+  // Connect
+  connect(_options, callback) {
+    var self = this;
+    if ('function' === typeof _options) (callback = _options), (_options = {});
+    if (_options == null) _options = this.s.clonedOptions;
+    if (!('function' === typeof callback)) callback = null;
+    _options = assign({}, this.s.clonedOptions, _options);
+    self.s.options = _options;
 
-  // Build default client information
-  clonedOptions.clientInfo = this.clientInfo;
-  // Do we have an application specific string
-  if (options.appname) {
-    clonedOptions.clientInfo.application = { name: options.appname };
+    // Update bufferMaxEntries
+    self.s.storeOptions.bufferMaxEntries =
+      typeof _options.bufferMaxEntries === 'number' ? _options.bufferMaxEntries : -1;
+
+    // Error handler
+    var connectErrorHandler = function() {
+      return function(err) {
+        // Remove all event handlers
+        var events = ['timeout', 'error', 'close'];
+        events.forEach(function(e) {
+          self.s.coreTopology.removeListener(e, connectHandlers[e]);
+        });
+
+        self.s.coreTopology.removeListener('connect', connectErrorHandler);
+
+        // Try to callback
+        try {
+          callback(err);
+        } catch (err) {
+          process.nextTick(function() {
+            throw err;
+          });
+        }
+      };
+    };
+
+    // Actual handler
+    var errorHandler = function(event) {
+      return function(err) {
+        if (event !== 'error') {
+          self.emit(event, err);
+        }
+      };
+    };
+
+    // Error handler
+    var reconnectHandler = function() {
+      self.emit('reconnect', self);
+      self.s.store.execute();
+    };
+
+    // Reconnect failed
+    var reconnectFailedHandler = function(err) {
+      self.emit('reconnectFailed', err);
+      self.s.store.flush(err);
+    };
+
+    // Destroy called on topology, perform cleanup
+    var destroyHandler = function() {
+      self.s.store.flush();
+    };
+
+    // relay the event
+    var relay = function(event) {
+      return function(t, server) {
+        self.emit(event, t, server);
+      };
+    };
+
+    // Connect handler
+    var connectHandler = function() {
+      // Clear out all the current handlers left over
+      ['timeout', 'error', 'close', 'destroy'].forEach(function(e) {
+        self.s.coreTopology.removeAllListeners(e);
+      });
+
+      // Set up listeners
+      self.s.coreTopology.on('timeout', errorHandler('timeout'));
+      self.s.coreTopology.once('error', errorHandler('error'));
+      self.s.coreTopology.on('close', errorHandler('close'));
+      // Only called on destroy
+      self.s.coreTopology.on('destroy', destroyHandler);
+
+      // Emit open event
+      self.emit('open', null, self);
+
+      // Return correctly
+      try {
+        callback(null, self);
+      } catch (err) {
+        console.log(err.stack);
+        process.nextTick(function() {
+          throw err;
+        });
+      }
+    };
+
+    // Set up listeners
+    var connectHandlers = {
+      timeout: connectErrorHandler('timeout'),
+      error: connectErrorHandler('error'),
+      close: connectErrorHandler('close')
+    };
+
+    // Clear out all the current handlers left over
+    [
+      'timeout',
+      'error',
+      'close',
+      'serverOpening',
+      'serverDescriptionChanged',
+      'serverHeartbeatStarted',
+      'serverHeartbeatSucceeded',
+      'serverHeartbeatFailed',
+      'serverClosed',
+      'topologyOpening',
+      'topologyClosed',
+      'topologyDescriptionChanged'
+    ].forEach(function(e) {
+      self.s.coreTopology.removeAllListeners(e);
+    });
+
+    // Add the event handlers
+    self.s.coreTopology.once('timeout', connectHandlers.timeout);
+    self.s.coreTopology.once('error', connectHandlers.error);
+    self.s.coreTopology.once('close', connectHandlers.close);
+    self.s.coreTopology.once('connect', connectHandler);
+    // Reconnect server
+    self.s.coreTopology.on('reconnect', reconnectHandler);
+    self.s.coreTopology.on('reconnectFailed', reconnectFailedHandler);
+
+    // Set up SDAM listeners
+    self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
+    self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
+    self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
+    self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
+    self.s.coreTopology.on('serverOpening', relay('serverOpening'));
+    self.s.coreTopology.on('serverClosed', relay('serverClosed'));
+    self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
+    self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
+    self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
+    self.s.coreTopology.on('attemptReconnect', relay('attemptReconnect'));
+    self.s.coreTopology.on('monitoring', relay('monitoring'));
+
+    // Start connection
+    self.s.coreTopology.connect(_options);
   }
+}
 
-  // Define the internal properties
-  this.s = {
-    // Create an instance of a server instance from mongodb-core
-    coreTopology: new CServer(clonedOptions),
-    // Server capabilities
-    sCapabilities: null,
-    // Cloned options
-    clonedOptions: clonedOptions,
-    // Reconnect
-    reconnect: clonedOptions.reconnect,
-    // Emit error
-    emitError: clonedOptions.emitError,
-    // Pool size
-    poolSize: clonedOptions.size,
-    // Store Options
-    storeOptions: storeOptions,
-    // Store
-    store: store,
-    // Host
-    host: host,
-    // Port
-    port: port,
-    // Options
-    options: options
-  };
-};
-
-inherits(Server, EventEmitter);
-
-var define = (Server.define = new Define('Server', Server, false));
-
-// BSON property
-Object.defineProperty(Server.prototype, 'bson', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.s.bson;
-  }
-});
-
-// Last ismaster
-Object.defineProperty(Server.prototype, 'isMasterDoc', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.lastIsMaster();
-  }
-});
-
-Object.defineProperty(Server.prototype, 'parserType', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.parserType;
-  }
-});
-
-// Last ismaster
 Object.defineProperty(Server.prototype, 'poolSize', {
   enumerable: true,
   get: function() {
@@ -288,272 +380,28 @@ Object.defineProperty(Server.prototype, 'port', {
   }
 });
 
-Object.defineProperty(Server.prototype, 'logicalSessionTimeoutMinutes', {
-  enumerable: true,
-  get: function() {
-    return this.s.coreTopology.logicalSessionTimeoutMinutes;
-  }
-});
-
-// Connect
-// Server.prototype.connect = function(db, _options, callback) {
-Server.prototype.connect = function(_options, callback) {
-  var self = this;
-  if ('function' === typeof _options) (callback = _options), (_options = {});
-  if (_options == null) _options = this.s.clonedOptions;
-  if (!('function' === typeof callback)) callback = null;
-  _options = assign({}, this.s.clonedOptions, _options);
-  self.s.options = _options;
-
-  // Update bufferMaxEntries
-  self.s.storeOptions.bufferMaxEntries =
-    typeof _options.bufferMaxEntries === 'number' ? _options.bufferMaxEntries : -1;
-
-  // Error handler
-  var connectErrorHandler = function() {
-    return function(err) {
-      // Remove all event handlers
-      var events = ['timeout', 'error', 'close'];
-      events.forEach(function(e) {
-        self.s.coreTopology.removeListener(e, connectHandlers[e]);
-      });
-
-      self.s.coreTopology.removeListener('connect', connectErrorHandler);
-
-      // Try to callback
-      try {
-        callback(err);
-      } catch (err) {
-        process.nextTick(function() {
-          throw err;
-        });
-      }
-    };
-  };
-
-  // Actual handler
-  var errorHandler = function(event) {
-    return function(err) {
-      if (event !== 'error') {
-        self.emit(event, err);
-      }
-    };
-  };
-
-  // Error handler
-  var reconnectHandler = function() {
-    self.emit('reconnect', self);
-    self.s.store.execute();
-  };
-
-  // Reconnect failed
-  var reconnectFailedHandler = function(err) {
-    self.emit('reconnectFailed', err);
-    self.s.store.flush(err);
-  };
-
-  // Destroy called on topology, perform cleanup
-  var destroyHandler = function() {
-    self.s.store.flush();
-  };
-
-  // relay the event
-  var relay = function(event) {
-    return function(t, server) {
-      self.emit(event, t, server);
-    };
-  };
-
-  // Connect handler
-  var connectHandler = function() {
-    // Clear out all the current handlers left over
-    ['timeout', 'error', 'close', 'destroy'].forEach(function(e) {
-      self.s.coreTopology.removeAllListeners(e);
-    });
-
-    // Set up listeners
-    self.s.coreTopology.on('timeout', errorHandler('timeout'));
-    self.s.coreTopology.once('error', errorHandler('error'));
-    self.s.coreTopology.on('close', errorHandler('close'));
-    // Only called on destroy
-    self.s.coreTopology.on('destroy', destroyHandler);
-
-    // Emit open event
-    self.emit('open', null, self);
-
-    // Return correctly
-    try {
-      callback(null, self);
-    } catch (err) {
-      console.log(err.stack);
-      process.nextTick(function() {
-        throw err;
-      });
-    }
-  };
-
-  // Set up listeners
-  var connectHandlers = {
-    timeout: connectErrorHandler('timeout'),
-    error: connectErrorHandler('error'),
-    close: connectErrorHandler('close')
-  };
-
-  // Clear out all the current handlers left over
-  [
-    'timeout',
-    'error',
-    'close',
-    'serverOpening',
-    'serverDescriptionChanged',
-    'serverHeartbeatStarted',
-    'serverHeartbeatSucceeded',
-    'serverHeartbeatFailed',
-    'serverClosed',
-    'topologyOpening',
-    'topologyClosed',
-    'topologyDescriptionChanged'
-  ].forEach(function(e) {
-    self.s.coreTopology.removeAllListeners(e);
-  });
-
-  // Add the event handlers
-  self.s.coreTopology.once('timeout', connectHandlers.timeout);
-  self.s.coreTopology.once('error', connectHandlers.error);
-  self.s.coreTopology.once('close', connectHandlers.close);
-  self.s.coreTopology.once('connect', connectHandler);
-  // Reconnect server
-  self.s.coreTopology.on('reconnect', reconnectHandler);
-  self.s.coreTopology.on('reconnectFailed', reconnectFailedHandler);
-
-  // Set up SDAM listeners
-  self.s.coreTopology.on('serverDescriptionChanged', relay('serverDescriptionChanged'));
-  self.s.coreTopology.on('serverHeartbeatStarted', relay('serverHeartbeatStarted'));
-  self.s.coreTopology.on('serverHeartbeatSucceeded', relay('serverHeartbeatSucceeded'));
-  self.s.coreTopology.on('serverHeartbeatFailed', relay('serverHeartbeatFailed'));
-  self.s.coreTopology.on('serverOpening', relay('serverOpening'));
-  self.s.coreTopology.on('serverClosed', relay('serverClosed'));
-  self.s.coreTopology.on('topologyOpening', relay('topologyOpening'));
-  self.s.coreTopology.on('topologyClosed', relay('topologyClosed'));
-  self.s.coreTopology.on('topologyDescriptionChanged', relay('topologyDescriptionChanged'));
-  self.s.coreTopology.on('attemptReconnect', relay('attemptReconnect'));
-  self.s.coreTopology.on('monitoring', relay('monitoring'));
-
-  // Start connection
-  self.s.coreTopology.connect(_options);
-};
-
-// Server capabilities
-Server.prototype.capabilities = function() {
-  if (this.s.sCapabilities) return this.s.sCapabilities;
-  if (this.s.coreTopology.lastIsMaster() == null) return null;
-  this.s.sCapabilities = new ServerCapabilities(this.s.coreTopology.lastIsMaster());
-  return this.s.sCapabilities;
-};
-
+const define = (Server.define = new Define('Server', Server, false));
 define.classMethod('capabilities', {
   callback: false,
   promise: false,
   returns: [ServerCapabilities]
 });
 
-// Command
-Server.prototype.command = function(ns, cmd, options, callback) {
-  this.s.coreTopology.command(ns, cmd, getReadPreference(options), callback);
-};
-
 define.classMethod('command', { callback: true, promise: false });
-
-// Insert
-Server.prototype.insert = function(ns, ops, options, callback) {
-  this.s.coreTopology.insert(ns, ops, options, callback);
-};
-
 define.classMethod('insert', { callback: true, promise: false });
-
-// Update
-Server.prototype.update = function(ns, ops, options, callback) {
-  this.s.coreTopology.update(ns, ops, options, callback);
-};
-
 define.classMethod('update', { callback: true, promise: false });
-
-// Remove
-Server.prototype.remove = function(ns, ops, options, callback) {
-  this.s.coreTopology.remove(ns, ops, options, callback);
-};
-
 define.classMethod('remove', { callback: true, promise: false });
-
-// IsConnected
-Server.prototype.isConnected = function() {
-  return this.s.coreTopology.isConnected();
-};
-
-Server.prototype.isDestroyed = function() {
-  return this.s.coreTopology.isDestroyed();
-};
-
 define.classMethod('isConnected', { callback: false, promise: false, returns: [Boolean] });
-
-// Insert
-Server.prototype.cursor = function(ns, cmd, options) {
-  options.disconnectHandler = this.s.store;
-  return this.s.coreTopology.cursor(ns, cmd, options);
-};
-
+define.classMethod('isDestroyed', { callback: false, promise: false, returns: [Boolean] });
 define.classMethod('cursor', {
   callback: false,
   promise: false,
   returns: [Cursor, AggregationCursor, CommandCursor]
 });
 
-Server.prototype.lastIsMaster = function() {
-  return this.s.coreTopology.lastIsMaster();
-};
-
-/**
- * Unref all sockets
- * @method
- */
-Server.prototype.unref = function() {
-  this.s.coreTopology.unref();
-};
-
-Server.prototype.close = function(forceClosed) {
-  this.s.coreTopology.destroy();
-  // We need to wash out all stored processes
-  if (forceClosed === true) {
-    this.s.storeOptions.force = forceClosed;
-    this.s.store.flush();
-  }
-};
-
 define.classMethod('close', { callback: false, promise: false });
-
-Server.prototype.auth = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.auth.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('auth', { callback: true, promise: false });
-
-Server.prototype.logout = function() {
-  var args = Array.prototype.slice.call(arguments, 0);
-  this.s.coreTopology.logout.apply(this.s.coreTopology, args);
-};
-
 define.classMethod('logout', { callback: true, promise: false });
-
-/**
- * All raw connections
- * @method
- * @return {array}
- */
-Server.prototype.connections = function() {
-  return this.s.coreTopology.connections();
-};
-
 define.classMethod('connections', { callback: false, promise: false, returns: [Array] });
 
 /**
