+++ lib/topologies/topology_base.js
@@ -1,7 +1,12 @@
 'use strict';
 
-var MongoError = require('mongodb-core').MongoError,
-  f = require('util').format;
+const EventEmitter = require('events'),
+  MongoError = require('mongodb-core').MongoError,
+  f = require('util').format,
+  os = require('os'),
+  getReadPreference = require('../utils').getReadPreference,
+  ReadPreference = require('../read_preference'),
+  CoreReadPreference = require('mongodb-core').ReadPreference;
 
 // The store of ops
 var Store = function(topology, storeOptions) {
@@ -256,5 +261,172 @@ var ServerCapabilities = function(ismaster) {
   setup_get_property(this, 'hasSessionSupport', sessionSupport);
 };
 
+// Get package.json variable
+const driverVersion = require('../../package.json').version,
+  nodejsversion = f('Node.js %s, %s', process.version, os.endianness()),
+  type = os.type(),
+  name = process.platform,
+  architecture = process.arch,
+  release = os.release();
+
+// Ensure the right read Preference object
+const translateReadPreference = function(options) {
+  if (typeof options.readPreference === 'string') {
+    options.readPreference = new CoreReadPreference(options.readPreference);
+  } else if (options.readPreference instanceof ReadPreference) {
+    options.readPreference = new CoreReadPreference(
+      options.readPreference.mode,
+      options.readPreference.tags,
+      { maxStalenessSeconds: options.readPreference.maxStalenessSeconds }
+    );
+  }
+
+  return options;
+};
+
+class TopologyBase extends EventEmitter {
+  constructor() {
+    super();
+
+    // Build default client information
+    this.clientInfo = {
+      driver: {
+        name: 'nodejs',
+        version: driverVersion
+      },
+      os: {
+        type: type,
+        name: name,
+        architecture: architecture,
+        version: release
+      },
+      platform: nodejsversion
+    };
+  }
+
+  // Server capabilities
+  capabilities() {
+    if (this.s.sCapabilities) return this.s.sCapabilities;
+    if (this.s.coreTopology.lastIsMaster() == null) return null;
+    this.s.sCapabilities = new ServerCapabilities(this.s.coreTopology.lastIsMaster());
+    return this.s.sCapabilities;
+  }
+
+  // Command
+  command(ns, cmd, options, callback) {
+    this.s.coreTopology.command(ns, cmd, getReadPreference(options), callback);
+  }
+
+  // Insert
+  insert(ns, ops, options, callback) {
+    this.s.coreTopology.insert(ns, ops, options, callback);
+  }
+
+  // Update
+  update(ns, ops, options, callback) {
+    this.s.coreTopology.update(ns, ops, options, callback);
+  }
+
+  // Remove
+  remove(ns, ops, options, callback) {
+    this.s.coreTopology.remove(ns, ops, options, callback);
+  }
+
+  // IsConnected
+  isConnected(options) {
+    options = options || {};
+    options = translateReadPreference(options);
+
+    return this.s.coreTopology.isConnected(options);
+  }
+
+  // IsDestroyed
+  isDestroyed() {
+    return this.s.coreTopology.isDestroyed();
+  }
+
+  // Cursor
+  cursor(ns, cmd, options) {
+    options = options || {};
+    options = translateReadPreference(options);
+    options.disconnectHandler = this.s.store;
+
+    return this.s.coreTopology.cursor(ns, cmd, options);
+  }
+
+  lastIsMaster() {
+    return this.s.coreTopology.lastIsMaster();
+  }
+
+  /**
+   * Unref all sockets
+   * @method
+   */
+  unref() {
+    return this.s.coreTopology.unref();
+  }
+
+  auth() {
+    var args = Array.prototype.slice.call(arguments, 0);
+    this.s.coreTopology.auth.apply(this.s.coreTopology, args);
+  }
+
+  logout() {
+    var args = Array.prototype.slice.call(arguments, 0);
+    this.s.coreTopology.logout.apply(this.s.coreTopology, args);
+  }
+
+  /**
+   * All raw connections
+   * @method
+   * @return {array}
+   */
+  connections() {
+    return this.s.coreTopology.connections();
+  }
+
+  close(forceClosed) {
+    this.s.coreTopology.destroy({
+      force: typeof forceClosed === 'boolean' ? forceClosed : false
+    });
+
+    // We need to wash out all stored processes
+    if (forceClosed === true) {
+      this.s.storeOptions.force = forceClosed;
+      this.s.store.flush();
+    }
+  }
+}
+
+// Properties
+Object.defineProperty(TopologyBase.prototype, 'isMasterDoc', {
+  enumerable: true,
+  get: function() {
+    return this.s.coreTopology.lastIsMaster();
+  }
+});
+
+Object.defineProperty(TopologyBase.prototype, 'bson', {
+  enumerable: true,
+  get: function() {
+    return this.s.coreTopology.s.bson;
+  }
+});
+
+Object.defineProperty(TopologyBase.prototype, 'parserType', {
+  enumerable: true,
+  get: function() {
+    return this.s.coreTopology.parserType;
+  }
+});
+
+Object.defineProperty(TopologyBase.prototype, 'logicalSessionTimeoutMinutes', {
+  enumerable: true,
+  get: function() {
+    return this.s.coreTopology.logicalSessionTimeoutMinutes;
+  }
+});
+
 exports.Store = Store;
 exports.ServerCapabilities = ServerCapabilities;
+exports.TopologyBase = TopologyBase;
